+++ views/admin/settings.php
@@ -81,6 +81,7 @@ $this->title = Yii::t('SpaceJoinQuestionsModule.base', 'Settings');
 
 
 
+
         <div class="form-group">
             <?= Button::primary(Yii::t('SpaceJoinQuestionsModule.base', 'Save Settings'))
                 ->submit()